+++ src_qt6/EasyReflectometryApp/Backends/Py/__init__.py
@@ -1,0 +1,3 @@
+from .py_backend import PyBackend
+
+__all__ = ['PyBackend']
