+++ src_qt6/EasyReflectometryApp/Backends/Py/py_backend.py
@@ -2,10 +2,10 @@ from PySide6.QtCore import QObject, Property
 
 from EasyApp.Logic.Logging import LoggerLevelHandler
 
-from .Py.home import Home
-from .Py.project import Project
-from .Py.status import Status
-from .Py.report import Report
+from .home import Home
+from .project import Project
+from .status import Status
+from .report import Report
 
 
 class PyBackend(QObject):
